requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Database
sqlite==3.42.0
sqlalchemy==2.0.19

# Visualization
matplotlib==3.7.2
plotly==5.17.0

# Web dashboard
flask==3.0.0
dash==2.14.0

# Utilities
pyyaml==6.0
joblib==1.3.2